Move event components to '/event-search' package
